+++ src/pipes/keys/keys.pipe.ts
@@ -4,8 +4,8 @@ import {Dictionary, isBlank} from '@jscrpt/common';
 /**
  * Gets object keys/property names
  */
-@Pipe({name: 'keys'})
-export class KeysPipe implements PipeTransform
+@Pipe({name: 'keys', standalone: true})
+export class KeysSAPipe implements PipeTransform
 {
     /**
      * Gets object keys/property names